lxc: copy version 2.0.4 from wr-kernel/meta-virt
This is the 'shipped' version with LIN9 and as such we default to use
this version via PREFERED_VERSION in the layer config.

The recipe and patches are 1:1 copy from wr-kernel/meta-virt. Even
though some patches are the same as those found in the 'files'
directory we keep copies in the 'lxc-2.0.4' directory (which will have
precedence) in order to allow for easier rebases.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>