openvswitch: create v2.6.1
This is a copy of the recipe, patches and other files needed to build
the version of OVS shipped with LIN9. Altough some patches were
identical to those found in the 'files' directory copies have been
made in the openvswitch-2.6.1 dir to make it easier to maintain the
2.6.1 version independently of the 2.7.0 version (which we know we
will be supporting in a future update of LIN9). This version is set as
the PREF VERSION to ensure LIN9 is built by default using the same
version originally shipped.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>