libvirt: do not defer postinstall to first boot when not necessary
Issue: LIN1018-2598

The postinstall invokes `/etc/init.d/populate-volatile.sh update'.
This is only useful for an on-target package upgrade. For now,
this postinstall in libvirt has already got a check against $D to
tell whether it's running on target or not, we do not need to use
pkg_postinstall_ontarget. Otherwise, we get unnecessary postinstall
deferred to first boot, which is not what we want.

Note that the postinstall was originally written as 'pkg_postinstall_${PN}',
but was incorrectly modified to be 'pkg_postinstall_ontarget_${PN}' by the
following commit.

  """
  commit 1e67a44db7ec87617455b3228ee5237ceb037173
  Author: He Zhe <zhe.he@windriver.com>
  Date:   Tue Jul 24 05:50:01 2018 -0700

      libvirt: Remove unsupported configure options and fix warning

  [snip]
  """

I think the warning about deferred postinstall was caused by some bbappend
file instead of the main recipe in this layer.

(LOCAL REV: NOT UPSTREAM) -- sent to meta-virtualization on 20181016

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>