openvswitch: Make openvswitch a machine_arch package
Openvswitch recipe depends on the kernel source, which is machine specific.
It can then use this machine specific resource to configure and build the
recipe.  This mix causes the hash values of this package to change based on
the machine, so this package must also be a machine_arch package.

(LOCAL REV: NOT UPSTREAM) -- Sent to meta-virtualization list on 20181017

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>